+++ src/Notifications/Entities/EntityFactory.php
@@ -8,9 +8,24 @@ class EntityFactory
 {
     public static function create(string $eventType, array $data): Entity
     {
+        // Map specific event entity types.
+        $eventEntityTypes = [
+            'payment_method.deleted' => DeletedPaymentMethod::class,
+        ];
+
+        $entity = $eventEntityTypes[$eventType] ?? self::resolveEntityClass($eventType);
+
+        return $entity::from($data);
+    }
+
+    /**
+     * @return class-string<Entity>
+     */
+    private static function resolveEntityClass(string $eventType): string
+    {
         $type = explode('.', $eventType);
-        $entity = $type[0] ?? 'Unknown';
-        $identifier = str_replace('_', '', ucwords(implode('_', $type), '_'));
+        $entity = self::snakeToPascalCase($type[0] ?? 'Unknown');
+        $identifier = self::snakeToPascalCase(implode('_', $type));
 
         /** @var class-string<Entity> $entity */
         $entity = sprintf('\Paddle\SDK\Notifications\Entities\%s', ucfirst($entity));
@@ -22,6 +37,11 @@ class EntityFactory
             throw new \UnexpectedValueException("Event type '{$identifier}' cannot be mapped to an object");
         }
 
-        return $entity::from($data);
+        return $entity;
+    }
+
+    private static function snakeToPascalCase(string $string): string
+    {
+        return str_replace('_', '', ucwords($string, '_'));
     }
 }